@@ -19,39 +19,30 @@ apply from: 'dependencies.gradle'
 apply from: 'common_methods.gradle'
 
 buildscript {
-  ext.kotlin_version = '1.2.31'
+  ext.kotlin_version = '1.3.50'
   repositories {
     jcenter()
     google()
   }
 
   dependencies {
-    classpath 'com.android.tools.build:gradle:3.1.0'
     classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-    classpath 'com.dicedmelon.gradle:jacoco-android:0.1.2'
-    classpath 'com.google.gms:google-services:3.1.0'
-    classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.4'
+    classpath 'com.android.tools.build:gradle:3.5.0'
+    classpath 'com.dicedmelon.gradle:jacoco-android:0.1.4'
     classpath 'com.github.dcendents:android-maven-gradle-plugin:2.0'
+    classpath 'com.google.gms:google-services:4.3.2'
+    classpath 'com.jfrog.bintray.gradle:gradle-bintray-plugin:1.4'
   }
 }
 
 allprojects {
-  ext {
-    compileSdkVersion = 27
-    buildToolsVersion = "27.0.3"
-    minSdkVersion = 16
-    targetSdkVersion = 27
-
-    supportLibVersion = "26.1.0"
-  }
 
   repositories {
-    jcenter()
     google()
+    jcenter()
     maven { url "http://jcenter.bintray.com" }
     maven { url 'https://dl.bintray.com/gigigo-desarrollo/maven' }
     maven { url 'https://jitpack.io' }
-    maven { url 'https://maven.google.com' }
   }
 
   tasks.withType(Javadoc).all {
@@ -69,7 +60,7 @@ task deployDebugLib(type: Exec) {
   workingDir "./"
   def command = "./gradlew"
   if (OperatingSystem.current().isWindows()) {
-    command = command + ".bat"
+    command = "gradlew.bat"
   }
 
   commandLine command, "install", "bintrayUpload", "-Pdebug"
@@ -79,7 +70,7 @@ task deployQualityLib(type: Exec) {
   workingDir "./"
   def command = "./gradlew"
   if (OperatingSystem.current().isWindows()) {
-    command = command + ".bat"
+    command = "gradlew.bat"
   }
 
   commandLine command, "install", "bintrayUpload", "-Pquality"
@@ -89,7 +80,7 @@ task deployLib(type: Exec) {
   workingDir "./"
   def command = "./gradlew"
   if (OperatingSystem.current().isWindows()) {
-    command = command + ".bat"
+    command = "gradlew.bat"
   }
 
   commandLine command, "install", "bintrayUpload"
