@@ -2,10 +2,10 @@ package com.gigigo.orchextrasdk.demo.test.util
 
 
 import android.app.Activity
-import android.support.test.espresso.Espresso.onView
-import android.support.test.espresso.UiController
-import android.support.test.espresso.ViewAction
-import android.support.test.espresso.matcher.ViewMatchers.isRoot
+import androidx.test.espresso.Espresso.onView
+import androidx.test.espresso.UiController
+import androidx.test.espresso.ViewAction
+import androidx.test.espresso.matcher.ViewMatchers.isRoot
 import android.view.View
 import com.squareup.spoon.Spoon
 import org.hamcrest.Matcher
