@@ -23,9 +23,9 @@ import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.os.Handler;
-import android.support.annotation.NonNull;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 import android.text.Editable;
 import android.text.TextWatcher;
 import android.view.View;
@@ -133,7 +133,6 @@ public class LoginActivity extends AppCompatActivity {
   }
 
   void loadProjectData() {
-
     apiKeyEditText.setText(projectDataList.get(currentProject).getApiKey());
     apiSecretEditText.setText(projectDataList.get(currentProject).getApiSecret());
   }
