@@ -9,15 +9,17 @@ import android.graphics.Color
 import android.media.RingtoneManager
 import android.os.Build.VERSION
 import android.os.Build.VERSION_CODES
-import android.support.annotation.RequiresApi
-import android.support.v4.app.NotificationCompat
+import androidx.annotation.RequiresApi
+import androidx.core.app.NotificationCompat
 import android.util.Log
 import com.gigigo.orchextra.core.R
 import com.gigigo.orchextra.core.data.datasources.network.models.ApiAction
 import com.gigigo.orchextra.core.data.datasources.network.models.toAction
 import com.gigigo.orchextra.core.domain.actions.ActionHandlerServiceExecutor
 import com.gigigo.orchextra.core.domain.datasources.DbDataSource
+import com.gigigo.orchextra.core.domain.datasources.NetworkDataSource
 import com.gigigo.orchextra.core.domain.entities.Action
+import com.gigigo.orchextra.core.domain.entities.TokenData
 import com.gigigo.orchextra.core.utils.LogUtils
 import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
@@ -26,108 +28,129 @@ import com.squareup.moshi.Moshi
 
 class OxFirebaseMessagingService : FirebaseMessagingService() {
 
-  private val moshi = Moshi.Builder().build()
-  private val apiActionAdapter = moshi.adapter(ApiAction::class.java)
-  private lateinit var actionHandlerServiceExecutor: ActionHandlerServiceExecutor
-  private lateinit var dbDataSource: DbDataSource
+    private val moshi = Moshi.Builder().build()
+    private val apiActionAdapter = moshi.adapter(ApiAction::class.java)
+    private lateinit var actionHandlerServiceExecutor: ActionHandlerServiceExecutor
+    private lateinit var dbDataSource: DbDataSource
 
-  override fun onCreate() {
-    super.onCreate()
-    actionHandlerServiceExecutor = ActionHandlerServiceExecutor.create(this)
-    dbDataSource = DbDataSource.create(baseContext)
-  }
-
-  override fun onMessageReceived(remoteMessage: RemoteMessage?) {
-
-    Log.d(TAG, "From: " + remoteMessage!!.from)
-
-    if (remoteMessage.data.isNotEmpty()) {
-      Log.d(TAG, "Message data payload: " + remoteMessage.data)
-
-
-      if (remoteMessage.data.containsKey("isOrchextra")
-          && remoteMessage.data["isOrchextra"] != "true") {
-        LogUtils.LOGD(TAG, "isOrchextra == false")
-        return
-      }
-
-      if (remoteMessage.data.containsKey("action")) {
-        val action = apiActionAdapter.fromJson(remoteMessage.data["action"]).toAction()
-        handleAction(action)
-      }
+    override fun onCreate() {
+        super.onCreate()
+        actionHandlerServiceExecutor = ActionHandlerServiceExecutor.create(this)
+        dbDataSource = DbDataSource.create(baseContext)
     }
 
-    if (remoteMessage.notification != null) {
-      Log.d(TAG, "Message Notification Body: " + remoteMessage.notification.body!!)
-      Log.d(TAG, "Message Notification Title: " + remoteMessage.notification.title!!)
+    override fun onNewToken(p0: String) {
+        super.onNewToken(p0)
 
-      sendNotification(
-          title = remoteMessage.notification.title ?: "Orchextra",
-          body = remoteMessage.notification.body ?: "")
+        val networkDataSource = NetworkDataSource.create(this)
+        val dbDataSource = DbDataSource.create(this)
+        dbDataSource.clearDevice()
+
+        val crm = dbDataSource.getCrm()
+        val device = dbDataSource.getDevice()
+        networkDataSource.updateTokenData(TokenData(crm = crm, device = device))
+
+        Log.w(TAG, "Refreshed token")
     }
-  }
+    override fun onMessageReceived(remoteMessage: RemoteMessage) {
 
-  private fun handleAction(action: Action) {
-    actionHandlerServiceExecutor.execute(action)
-  }
+        Log.d(TAG, "From: ${remoteMessage.from}")
 
-  private fun sendNotification(title: String, body: String) {
+        if (remoteMessage.data.isNotEmpty()) {
+            Log.d(TAG, "Message data payload: " + remoteMessage.data)
 
-    if (VERSION.SDK_INT >= VERSION_CODES.O) {
-      createNotificationChannel()
+            if (remoteMessage.data.containsKey("isOrchextra")
+                && remoteMessage.data["isOrchextra"] != "true"
+            ) {
+                LogUtils.LOGD(TAG, "isOrchextra == false")
+                return
+            }
+
+            if (remoteMessage.data.containsKey("action")) {
+                remoteMessage.data["action"]?.let {
+                    apiActionAdapter.fromJson(it)?.toAction()?.let { action ->
+                        handleAction(action)
+                    }
+                }
+            }
+        }
+
+        remoteMessage.notification?.let { notification ->
+            Log.d(TAG, "Message Notification Body: ${notification.body}")
+            Log.d(TAG, "Message Notification Title: ${notification.title}")
+
+            sendNotification(
+                title = notification.title ?: "Orchextra",
+                body = notification.body ?: ""
+            )
+        }
     }
 
-    val defaultSoundUri = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION)
-    val notificationBuilder = NotificationCompat.Builder(this, CHANNEL_ID)
-        .setSmallIcon(R.drawable.ox_close)
-        .setContentTitle(title)
-        .setContentText(body)
-        .setAutoCancel(true)
-        .setSound(defaultSoundUri)
-
-    val notificationActivityName = dbDataSource.getNotificationActivityName()
-    if (notificationActivityName.isNotEmpty()) {
-      if (getNotificationActivityClass(notificationActivityName) != null) {
-        val intent = Intent(this, getNotificationActivityClass(notificationActivityName))
-        intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
-        val pendingIntent = PendingIntent.getActivity(this, 0, intent,
-            PendingIntent.FLAG_ONE_SHOT)
-
-        notificationBuilder.setContentIntent(pendingIntent)
-      }
+    private fun handleAction(action: Action) {
+        actionHandlerServiceExecutor.execute(action)
     }
 
-    val notificationManager = getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
-    notificationManager.notify(0x9387, notificationBuilder.build())
-  }
+    private fun sendNotification(title: String, body: String) {
 
-  @RequiresApi(VERSION_CODES.O)
-  private fun createNotificationChannel() {
+        if (VERSION.SDK_INT >= VERSION_CODES.O) {
+            createNotificationChannel()
+        }
 
-    val mNotificationManager = getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+        val defaultSoundUri = RingtoneManager.getDefaultUri(RingtoneManager.TYPE_NOTIFICATION)
+        val notificationBuilder = NotificationCompat.Builder(this, CHANNEL_ID)
+            .setSmallIcon(R.drawable.ox_close)
+            .setContentTitle(title)
+            .setContentText(body)
+            .setAutoCancel(true)
+            .setSound(defaultSoundUri)
 
-    val name = getString(R.string.app_name)
-    val description = getString(R.string.app_name)
-    val importance = NotificationManager.IMPORTANCE_DEFAULT
-    val mChannel = NotificationChannel(CHANNEL_ID, name, importance)
+        val notificationActivityName = dbDataSource.getNotificationActivityName()
+        if (notificationActivityName.isNotEmpty()) {
+            if (getNotificationActivityClass(notificationActivityName) != null) {
+                val intent = Intent(this, getNotificationActivityClass(notificationActivityName))
+                intent.addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
+                val pendingIntent = PendingIntent.getActivity(
+                    this, 0, intent,
+                    PendingIntent.FLAG_ONE_SHOT
+                )
 
-    mChannel.description = description
-    mChannel.enableLights(true)
+                notificationBuilder.setContentIntent(pendingIntent)
+            }
+        }
 
-    mChannel.lightColor = Color.RED
-    mChannel.enableVibration(true)
-    mChannel.vibrationPattern = longArrayOf(100, 200, 300, 400, 500, 400, 300, 200, 400)
-    mNotificationManager.createNotificationChannel(mChannel)
-  }
+        val notificationManager =
+            getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+        notificationManager.notify(0x9387, notificationBuilder.build())
+    }
 
-  private fun getNotificationActivityClass(activityName: String): Class<*>? = try {
-    Class.forName(activityName)
-  } catch (exception: ClassNotFoundException) {
-    null
-  }
+    @RequiresApi(VERSION_CODES.O)
+    private fun createNotificationChannel() {
 
-  companion object {
-    private val TAG = "OxFirebaseMsgService"
-    private val CHANNEL_ID = "ox_push_notification"
-  }
+        val mNotificationManager =
+            getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
+
+        val name = getString(R.string.app_name)
+        val description = getString(R.string.app_name)
+        val importance = NotificationManager.IMPORTANCE_DEFAULT
+        val mChannel = NotificationChannel(CHANNEL_ID, name, importance)
+
+        mChannel.description = description
+        mChannel.enableLights(true)
+
+        mChannel.lightColor = Color.RED
+        mChannel.enableVibration(true)
+        mChannel.vibrationPattern = longArrayOf(100, 200, 300, 400, 500, 400, 300, 200, 400)
+        mNotificationManager.createNotificationChannel(mChannel)
+    }
+
+    private fun getNotificationActivityClass(activityName: String): Class<*>? = try {
+        Class.forName(activityName)
+    } catch (exception: ClassNotFoundException) {
+        null
+    }
+
+    companion object {
+        private const val TAG = "OxFirebaseMsgService"
+        private const val CHANNEL_ID = "ox_push_notification"
+    }
 }
