@@ -25,12 +25,11 @@ import android.content.Intent
 import android.content.pm.PackageManager
 import android.os.Bundle
 import android.preference.PreferenceManager
-import android.support.v4.content.ContextCompat
+import androidx.core.content.ContextCompat
 import com.gigigo.orchextra.core.domain.entities.GeoMarketing
 import com.gigigo.orchextra.core.domain.triggers.OxTrigger
 import com.gigigo.orchextra.core.utils.LogUtils
 import com.gigigo.orchextra.core.utils.LogUtils.LOGD
-import com.gigigo.orchextra.core.utils.LogUtils.LOGE
 import com.gigigo.orchextra.geofence.utils.toGeofence
 import com.google.android.gms.common.api.GoogleApiClient
 import com.google.android.gms.location.Geofence
@@ -41,124 +40,125 @@ import com.google.android.gms.tasks.OnCompleteListener
 import com.google.android.gms.tasks.Task
 
 
-class OxGeofenceImp(private val context: Application,
-    private val geofencingClient: GeofencingClient) : OxTrigger<List<GeoMarketing>>, OnCompleteListener<Void> {
+class OxGeofenceImp private constructor(
+    private val context: Application,
+    private val geofencingClient: GeofencingClient
+) : OxTrigger<List<GeoMarketing>>, OnCompleteListener<Void> {
 
-  private val TAG = LogUtils.makeLogTag(OxGeofenceImp::class.java)
-  private var geofenceList: List<Geofence> = ArrayList()
-  private var geofencePendingIntent: PendingIntent? = null
+    private var geofenceList: List<Geofence> = ArrayList()
+    private var geofencePendingIntent: PendingIntent? = null
 
-  override fun init() {
-
-    if (geofenceList.isEmpty()) {
-      throw IllegalArgumentException("Empty geofence list")
+    override fun init() {
+        if (geofenceList.isEmpty()) {
+            LOGD(TAG, "geofences list is empty")
+            return
+        }
+        connectWithCallbacks()
+        addGeofences()
     }
 
-    connectWithCallbacks()
-    addGeofences()
-  }
-
-  override fun setConfig(config: List<GeoMarketing>) {
-    geofenceList = config.map { it.toGeofence() }
-  }
-
-  override fun finish() {
-    removeGeofences()
-  }
-
-  private fun addGeofences() {
-    if (ContextCompat.checkSelfPermission(context,
-            Manifest.permission.ACCESS_FINE_LOCATION) == PackageManager.PERMISSION_GRANTED) {
-      geofencingClient.addGeofences(getGeofencingRequest(), getGeofencePendingIntent())
-          .addOnCompleteListener(this)
-    } else {
-      throw SecurityException("Geofence trigger needs ACCESS_FINE_LOCATION permission")
-    }
-  }
-
-  private fun removeGeofences() {
-    geofencingClient.removeGeofences(getGeofencePendingIntent()).addOnCompleteListener(this)
-  }
-
-  private fun connectWithCallbacks() {
-    val googleApiClient = GoogleApiClient.Builder(context)
-        .addApi(LocationServices.API)
-        .addConnectionCallbacks(connectionAddListener)
-        .addOnConnectionFailedListener(connectionFailedListener)
-        .build()
-    googleApiClient.connect()
-  }
-
-  private val connectionAddListener = object : GoogleApiClient.ConnectionCallbacks {
-    override fun onConnected(bundle: Bundle?) {
-      print("connectionAddListener")
+    override fun setConfig(config: List<GeoMarketing>) {
+        geofenceList = config.map { it.toGeofence() }
     }
 
-    override fun onConnectionSuspended(i: Int) {
-      print("onConnectionSuspended")
-    }
-  }
-
-  private val connectionFailedListener = GoogleApiClient.OnConnectionFailedListener {
-    print("connectionFailedListener")
-  }
-
-  override fun onComplete(task: Task<Void>) {
-    if (task.isSuccessful) {
-      updateGeofencesAdded(!getGeofencesAdded())
-
-      val messageId = if (getGeofencesAdded()) {
-        R.string.geofences_added
-      } else {
-        R.string.geofences_removed
-      }
-      LOGD(TAG, "onComplete: $messageId")
-    } else {
-
-      val errorMessage = GeofenceErrorMessages.getErrorString(context, task.exception)
-      LOGE(TAG, "onComplete: $errorMessage")
-    }
-  }
-
-  private fun getGeofencePendingIntent(): PendingIntent {
-
-    if (geofencePendingIntent != null) {
-      return geofencePendingIntent as PendingIntent
+    override fun finish() {
+        removeGeofences()
     }
 
-    val intent = Intent(context, GeofenceTransitionsIntentService::class.java)
-    geofencePendingIntent = PendingIntent.getService(context, 0, intent,
-        PendingIntent.FLAG_UPDATE_CURRENT)
+    private fun addGeofences() {
+        if (
+            ContextCompat.checkSelfPermission(context, Manifest.permission.ACCESS_FINE_LOCATION)
+            == PackageManager.PERMISSION_GRANTED
+        ) {
+            geofencingClient.addGeofences(getGeofencingRequest(), getGeofencePendingIntent())
+                .addOnCompleteListener(this)
+        } else {
+            throw SecurityException("Geofence trigger needs ACCESS_FINE_LOCATION permission")
+        }
+    }
 
-    return geofencePendingIntent as PendingIntent
-  }
+    private fun removeGeofences() {
+        geofencingClient.removeGeofences(getGeofencePendingIntent()).addOnCompleteListener(this)
+    }
 
-  private fun getGeofencingRequest(): GeofencingRequest {
-    val builder = GeofencingRequest.Builder()
-    builder.setInitialTrigger(
-        GeofencingRequest.INITIAL_TRIGGER_ENTER
-            or GeofencingRequest.INITIAL_TRIGGER_EXIT
-            or GeofencingRequest.INITIAL_TRIGGER_DWELL)
-    builder.addGeofences(geofenceList)
-    return builder.build()
-  }
+    private fun getGeofencePendingIntent(): PendingIntent {
+        if (geofencePendingIntent != null) {
+            return geofencePendingIntent as PendingIntent
+        }
 
-  private fun getGeofencesAdded(): Boolean {
-    return PreferenceManager.getDefaultSharedPreferences(context).getBoolean(
-        GEOFENCES_ADDED_KEY, false)
-  }
+        val intent = Intent(context, GeofenceBroadcastReceiver::class.java)
+        geofencePendingIntent = PendingIntent.getBroadcast(
+            context, 0, intent,
+            PendingIntent.FLAG_UPDATE_CURRENT
+        )
 
-  private fun updateGeofencesAdded(added: Boolean) {
-    PreferenceManager.getDefaultSharedPreferences(context)
-        .edit()
-        .putBoolean(GEOFENCES_ADDED_KEY, added)
-        .apply()
-  }
+        return geofencePendingIntent as PendingIntent
+    }
 
-  companion object Factory {
-    private const val GEOFENCES_ADDED_KEY = "GEOFENCES_ADDED_KEY"
+    private fun connectWithCallbacks() {
+        val googleApiClient = GoogleApiClient.Builder(context)
+            .addApi(LocationServices.API)
+            .addConnectionCallbacks(connectionAddListener)
+            .addOnConnectionFailedListener(connectionFailedListener)
+            .build()
+        googleApiClient.connect()
+    }
 
-    fun create(context: Application): OxGeofenceImp = OxGeofenceImp(context,
-        LocationServices.getGeofencingClient(context))
-  }
+    private val connectionAddListener = object : GoogleApiClient.ConnectionCallbacks {
+        override fun onConnected(bundle: Bundle?) {
+            LOGD(TAG, "connectionAddListener")
+        }
+
+        override fun onConnectionSuspended(i: Int) {
+            LOGD(TAG, "onConnectionSuspended")
+        }
+    }
+
+    private val connectionFailedListener = GoogleApiClient.OnConnectionFailedListener {
+        LogUtils.LOGE(TAG, "connectionFailedListener")
+    }
+
+    override fun onComplete(task: Task<Void>) {
+        if (task.isSuccessful) {
+            updateGeofencesAdded(!getGeofencesAdded())
+
+            val messageId =
+                if (getGeofencesAdded()) R.string.geofences_added
+                else R.string.geofences_removed
+            LOGD(TAG, "onComplete: $messageId")
+        } else {
+            task.exception?.let {
+                LogUtils.LOGE(TAG, it.message ?: "no exception message")
+            }
+            val errorMessage = GeofenceErrorMessages.getErrorString(context, task.exception)
+            LogUtils.LOGE(TAG, "onComplete: $errorMessage")
+        }
+    }
+
+    private fun getGeofencingRequest(): GeofencingRequest {
+        val builder = GeofencingRequest.Builder()
+        builder.setInitialTrigger(GeofencingRequest.INITIAL_TRIGGER_ENTER)
+        builder.addGeofences(geofenceList)
+        return builder.build()
+    }
+
+    private fun getGeofencesAdded(): Boolean {
+        return PreferenceManager.getDefaultSharedPreferences(context)
+            .getBoolean(GEOFENCES_ADDED_KEY, false)
+    }
+
+    private fun updateGeofencesAdded(added: Boolean) {
+        PreferenceManager.getDefaultSharedPreferences(context)
+            .edit()
+            .putBoolean(GEOFENCES_ADDED_KEY, added)
+            .apply()
+    }
+
+    companion object Factory {
+        private val TAG = LogUtils.makeLogTag(OxGeofenceImp::class.java)
+        private const val GEOFENCES_ADDED_KEY = "GEOFENCES_ADDED_KEY"
+
+        fun create(context: Application) =
+            OxGeofenceImp(context, LocationServices.getGeofencingClient(context))
+    }
 }
