@@ -19,36 +19,86 @@
 package com.gigigo.orchextra.core.domain.triggers
 
 import android.app.IntentService
+import android.app.NotificationChannel
+import android.app.NotificationManager
 import android.content.Context
 import android.content.Intent
+import android.graphics.Color
 import android.os.Build.VERSION
 import android.os.Build.VERSION_CODES
+import androidx.core.app.NotificationCompat
+import android.util.Log
+import com.gigigo.orchextra.core.Orchextra.NOTIFICATION_CHANNEL
+import com.gigigo.orchextra.core.R
+import com.gigigo.orchextra.core.R.string
 import com.gigigo.orchextra.core.domain.entities.Trigger
 import com.gigigo.orchextra.core.utils.LogUtils
 
 class TriggerHandlerService : IntentService(TAG) {
 
-  override fun onHandleIntent(intent: Intent) {
+    override fun onCreate() {
+        super.onCreate()
+        val manager = getSystemService(Context.NOTIFICATION_SERVICE) as NotificationManager
 
-    val triggerListener: TriggerListener = TriggerManager.create(this)
-    val trigger = intent.getParcelableExtra<Trigger>(TRIGGER_EXTRA)
-    triggerListener.onTriggerDetected(trigger)
-  }
-
-  companion object Navigator {
-    private val TAG = LogUtils.makeLogTag(TriggerHandlerService::class.java)
-    private const val TRIGGER_EXTRA = "trigger_extra"
-
-    fun start(context: Context, trigger: Trigger) {
-      val intent = Intent(context, TriggerHandlerService::class.java)
-      intent.putExtra(TRIGGER_EXTRA, trigger)
-      try {
-        context.startService(intent)
-      } catch (exception: IllegalStateException) {
         if (VERSION.SDK_INT >= VERSION_CODES.O) {
-          context.startForegroundService(intent)
+            val channel = NotificationChannel(
+                NOTIFICATION_CHANNEL, getString(string.app_name),
+                NotificationManager.IMPORTANCE_LOW
+            ).apply {
+                lightColor = Color.RED
+                lockscreenVisibility = android.app.Notification.VISIBILITY_PRIVATE
+                enableVibration(false)
+                setSound(null, null)
+            }
+            manager.createNotificationChannel(channel)
         }
-      }
+
+        val notificationBuilder = NotificationCompat.Builder(this, NOTIFICATION_CHANNEL)
+            .setSmallIcon(R.drawable.ox_notification_alpha_small_icon)
+            .setContentTitle(getString(R.string.app_name))
+            .setPriority(NotificationCompat.PRIORITY_LOW)
+            .setVibrate(longArrayOf(0, 0, 0))
+            .setLights(Color.RED, 0, 100)
+            .setSound(null)
+
+        startForeground(NOTIFICATION_ID, notificationBuilder.build())
     }
-  }
+
+    override fun onDestroy() {
+        super.onDestroy()
+
+        val notificationManager = applicationContext.getSystemService(
+            Context.NOTIFICATION_SERVICE
+        ) as NotificationManager
+        notificationManager.cancel(NOTIFICATION_ID)
+    }
+
+    override fun onHandleIntent(intent: Intent?) {
+
+        val triggerListener: TriggerListener = TriggerManager.create(this)
+        val trigger = intent?.getParcelableExtra<Trigger>(TRIGGER_EXTRA)
+        if (trigger == null) {
+            Log.e(TAG, "trigger empty")
+            return
+        }
+        triggerListener.onTriggerDetected(trigger)
+    }
+
+    companion object Navigator {
+        private val TAG = LogUtils.makeLogTag(TriggerHandlerService::class.java)
+        private const val TRIGGER_EXTRA = "trigger_extra"
+        private const val NOTIFICATION_ID = 0x654
+
+        fun start(context: Context, trigger: Trigger) {
+            val intent = Intent(context, TriggerHandlerService::class.java)
+            intent.putExtra(TRIGGER_EXTRA, trigger)
+            try {
+                context.startService(intent)
+            } catch (exception: IllegalStateException) {
+                if (VERSION.SDK_INT >= VERSION_CODES.O) {
+                    context.startForegroundService(intent)
+                }
+            }
+        }
+    }
 }
